refactor: new sub-reconciler for namespace reconciliation
