+++ internal/controllers/addon/namespace_reconciler_test.go
@@ -2,6 +2,7 @@ package addon
 
 import (
 	"context"
+	"errors"
 	"testing"
 
 	"github.com/stretchr/testify/assert"
@@ -21,10 +22,9 @@ import (
 func TestEnsureWantedNamespaces_AddonWithoutNamespaces(t *testing.T) {
 	c := testutil.NewClient()
 
-	r := &AddonReconciler{
-		Client: c,
-		Log:    testutil.NewLogger(t),
-		Scheme: testutil.NewTestSchemeWithAddonsv1alpha1(),
+	r := &namespaceReconciler{
+		scheme: testutil.NewTestSchemeWithAddonsv1alpha1(),
+		client: c,
 	}
 
 	ctx := context.Background()
@@ -40,10 +40,9 @@ func TestEnsureWantedNamespaces_AddonWithSingleNamespace_Collision(t *testing.T)
 		arg := args.Get(2).(*corev1.Namespace)
 		testutil.NewTestExistingNamespace().DeepCopyInto(arg)
 	}).Return(nil)
-	r := &AddonReconciler{
-		Client: c,
-		Log:    testutil.NewLogger(t),
-		Scheme: testutil.NewTestSchemeWithAddonsv1alpha1(),
+	r := &namespaceReconciler{
+		scheme: testutil.NewTestSchemeWithAddonsv1alpha1(),
+		client: c,
 	}
 
 	ctx := context.Background()
@@ -72,10 +71,9 @@ func TestEnsureWantedNamespaces_AddonWithSingleNamespace_NoCollision(t *testing.
 		}
 	}).Return(nil)
 
-	r := &AddonReconciler{
-		Client: c,
-		Log:    testutil.NewLogger(t),
-		Scheme: testutil.NewTestSchemeWithAddonsv1alpha1(),
+	r := &namespaceReconciler{
+		scheme: testutil.NewTestSchemeWithAddonsv1alpha1(),
+		client: c,
 	}
 
 	ctx := context.Background()
@@ -97,10 +95,9 @@ func TestEnsureWantedNamespaces_AddonWithMultipleNamespaces_NoCollision(t *testi
 		}
 	}).Return(nil)
 
-	r := &AddonReconciler{
-		Client: c,
-		Log:    testutil.NewLogger(t),
-		Scheme: testutil.NewTestSchemeWithAddonsv1alpha1(),
+	r := &namespaceReconciler{
+		scheme: testutil.NewTestSchemeWithAddonsv1alpha1(),
+		client: c,
 	}
 
 	ctx := context.Background()
@@ -135,10 +132,9 @@ func TestEnsureWantedNamespaces_AddonWithMultipleNamespaces_SingleCollision(t *t
 		}).
 		Return(nil)
 
-	r := &AddonReconciler{
-		Client: c,
-		Log:    testutil.NewLogger(t),
-		Scheme: testutil.NewTestSchemeWithAddonsv1alpha1(),
+	r := &namespaceReconciler{
+		scheme: testutil.NewTestSchemeWithAddonsv1alpha1(),
+		client: c,
 	}
 
 	ctx := context.Background()
@@ -165,10 +161,9 @@ func TestEnsureWantedNamespaces_AddonWithMultipleNamespaces_MultipleCollisions(t
 		}).
 		Return(nil)
 
-	r := &AddonReconciler{
-		Client: c,
-		Log:    testutil.NewLogger(t),
-		Scheme: testutil.NewTestSchemeWithAddonsv1alpha1(),
+	r := &namespaceReconciler{
+		scheme: testutil.NewTestSchemeWithAddonsv1alpha1(),
+		client: c,
 	}
 
 	ctx := context.Background()
@@ -194,10 +189,9 @@ func TestEnsureNamespace_Create(t *testing.T) {
 	c.On("Get", testutil.IsContext, testutil.IsObjectKey, testutil.IsCoreV1NamespacePtr).Return(testutil.NewTestErrNotFound())
 	c.On("Create", testutil.IsContext, testutil.IsCoreV1NamespacePtr, mock.Anything).Return(nil)
 
-	r := &AddonReconciler{
-		Client: c,
-		Log:    testutil.NewLogger(t),
-		Scheme: testutil.NewTestSchemeWithAddonsv1alpha1(),
+	r := &namespaceReconciler{
+		scheme: testutil.NewTestSchemeWithAddonsv1alpha1(),
+		client: c,
 	}
 
 	ctx := context.Background()
@@ -220,10 +214,9 @@ func TestEnsureNamespace_CreateWithLabels(t *testing.T) {
 		}).
 		Return(nil)
 
-	r := &AddonReconciler{
-		Client: c,
-		Log:    testutil.NewLogger(t),
-		Scheme: testutil.NewTestSchemeWithAddonsv1alpha1(),
+	r := &namespaceReconciler{
+		scheme: testutil.NewTestSchemeWithAddonsv1alpha1(),
+		client: c,
 	}
 
 	ctx := context.Background()
@@ -375,4 +368,124 @@ func TestReconcileNamespace_CreateWithClientError(t *testing.T) {
 	c.AssertCalled(t, "Get", testutil.IsContext, client.ObjectKey{
 		Name: "namespace-1",
 	}, testutil.IsCoreV1NamespacePtr)
+}
+
+func TestEnsureDeletionOfUnwantedNamespaces_NoNamespacesInSpec_NoNamespacesInCluster(t *testing.T) {
+	c := testutil.NewClient()
+
+	c.On("List", testutil.IsContext, testutil.IsCoreV1NamespaceListPtr, mock.Anything).
+		Return(nil)
+
+	r := &namespaceReconciler{
+		scheme: testutil.NewTestSchemeWithAddonsv1alpha1(),
+		client: c,
+	}
+
+	ctx := context.Background()
+	err := r.ensureDeletionOfUnwantedNamespaces(ctx, testutil.NewTestAddonWithoutNamespace())
+	require.NoError(t, err)
+	c.AssertExpectations(t)
+}
+
+func TestEnsureDeletionOfUnwantedNamespaces_NoNamespacesInSpec_NamespaceInCluster(t *testing.T) {
+	c := testutil.NewClient()
+
+	addon := testutil.NewTestAddonWithoutNamespace()
+	existingNamespace := testutil.NewTestNamespace()
+
+	c.On("List", testutil.IsContext, testutil.IsCoreV1NamespaceListPtr, mock.Anything).
+		Run(func(args mock.Arguments) {
+			arg := args.Get(1).(*corev1.NamespaceList)
+			namespaceList := corev1.NamespaceList{
+				Items: []corev1.Namespace{
+					*existingNamespace,
+				},
+			}
+			namespaceList.DeepCopyInto(arg)
+		}).
+		Return(nil)
+	c.On("Delete", testutil.IsContext, testutil.IsCoreV1NamespacePtr, mock.Anything).
+		Return(nil)
+
+	r := &namespaceReconciler{
+		scheme: testutil.NewTestSchemeWithAddonsv1alpha1(),
+		client: c,
+	}
+
+	ctx := context.Background()
+	err := r.ensureDeletionOfUnwantedNamespaces(ctx, addon)
+	require.NoError(t, err)
+	c.AssertExpectations(t)
+	c.AssertCalled(t, "List", testutil.IsContext,
+		testutil.IsCoreV1NamespaceListPtr,
+		// verify that the list call did use the correct labelSelector
+		mock.MatchedBy(func(listOptions []client.ListOption) bool {
+			testListOptions := &client.ListOptions{}
+			listOptions[0].ApplyToList(testListOptions)
+			testLabelSelectorString := testListOptions.LabelSelector.String()
+			return len(testLabelSelectorString) > 0 &&
+				testLabelSelectorString == controllers.CommonLabelsAsLabelSelector(addon).String()
+		}))
+	c.AssertCalled(t, "Delete", testutil.IsContext,
+		mock.MatchedBy(func(val *corev1.Namespace) bool {
+			return val.Name == existingNamespace.Name
+		}),
+		mock.Anything)
+}
+
+func TestEnsureDeletionOfUnwantedNamespaces_NamespacesInSpec_matching_NamespacesInCluster(t *testing.T) {
+	c := testutil.NewClient()
+
+	addon := testutil.NewTestAddonWithSingleNamespace()
+	existingNamespace := testutil.NewTestNamespace()
+
+	c.On("List", testutil.IsContext, testutil.IsCoreV1NamespaceListPtr, mock.Anything).
+		Run(func(args mock.Arguments) {
+			arg := args.Get(1).(*corev1.NamespaceList)
+			namespaceList := corev1.NamespaceList{
+				Items: []corev1.Namespace{
+					*existingNamespace,
+				},
+			}
+			namespaceList.DeepCopyInto(arg)
+		}).
+		Return(nil)
+
+	r := &namespaceReconciler{
+		scheme: testutil.NewTestSchemeWithAddonsv1alpha1(),
+		client: c,
+	}
+
+	ctx := context.Background()
+	err := r.ensureDeletionOfUnwantedNamespaces(ctx, addon)
+	require.NoError(t, err)
+	c.AssertExpectations(t)
+	c.AssertCalled(t, "List", testutil.IsContext,
+		testutil.IsCoreV1NamespaceListPtr,
+		// verify that the list call did use the correct labelSelector
+		mock.MatchedBy(func(listOptions []client.ListOption) bool {
+			testListOptions := &client.ListOptions{}
+			listOptions[0].ApplyToList(testListOptions)
+			testLabelSelectorString := testListOptions.LabelSelector.String()
+			return len(testLabelSelectorString) > 0 &&
+				testLabelSelectorString == controllers.CommonLabelsAsLabelSelector(addon).String()
+		}))
+}
+
+func TestEnsureDeletionOfUnwantedNamespaces_NoNamespacesInSpec_WithClientError(t *testing.T) {
+	timeoutErr := k8sApiErrors.NewTimeoutError("for testing", 1)
+
+	c := testutil.NewClient()
+	c.On("List", testutil.IsContext, testutil.IsCoreV1NamespaceListPtr, mock.Anything).
+		Return(timeoutErr)
+
+	r := &namespaceReconciler{
+		scheme: testutil.NewTestSchemeWithAddonsv1alpha1(),
+		client: c,
+	}
+
+	ctx := context.Background()
+	err := r.ensureDeletionOfUnwantedNamespaces(ctx, testutil.NewTestAddonWithoutNamespace())
+	require.EqualError(t, errors.Unwrap(err), timeoutErr.Error())
+	c.AssertExpectations(t)
 }
